♻️ Refactor :  types 관심사별 분리 및 경로변경
