+++ client/components/common/Header/index.tsx
@@ -8,10 +8,10 @@ import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
 
 import { useEffect } from 'react';
 import { client } from '../../../libs/client';
-import { IDataHeader } from '../../../interfaces';
 import { dataHeaderAtom } from '../../../atoms/userAtom';
 import { isLoginAtom } from '../../../atoms/loginAtom';
 import { renderingAtom } from '../../../atoms/renderingAtom';
+import { IDataHeader } from '../../../types/user';
 
 export const Header = () => {
   const [isLogin, setIsLogin] = useRecoilState(isLoginAtom);
