+++ client/components/questions/question-detail/AnswerContent/AnswerContainer.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useRouter } from 'next/router';
 import { useFetch } from '../../../../libs/useFetchSWR';
-import { Answer } from '../../../../libs/interfaces';
+import { Answer } from '../../../../types/answer';
 import { AnswerContent, AnswerProps } from './index';
 
 type AnswerListProps = {
