+++ client/types/calendar.ts
@@ -1,0 +1,9 @@
+export interface ICalendar {
+  jobId: number;
+  name: string;
+  state: string;
+  careerRequirement: string;
+  url: string;
+  startDate: string;
+  endDate: string;
+}
