+++ client/types/comment.ts
@@ -1,0 +1,13 @@
+import { UserInfo } from './user';
+import { Avatar } from './user';
+
+export interface CommentResp {
+  answerId?: number | null;
+  commentId: number;
+  articleId: number;
+  content: string;
+  createdAt: string;
+  lastModifiedAt: string;
+  userInfo: UserInfo;
+  avatar: Avatar;
+}
