+++ client/types/login.ts
@@ -1,0 +1,10 @@
+export interface AuthResp {
+  email: string;
+  authKey?: string;
+}
+
+export type DecodedResp = {
+  sub: string;
+  id: number;
+  nickname: string;
+};
