+++ client/types/pageInfo.ts
@@ -1,0 +1,6 @@
+export interface PageInfo {
+  page: number;
+  size: number;
+  totalElements: number;
+  sort: { empty: boolean; unsorted: boolean; sorted: boolean };
+}
